reorg and removed code for attest-blob
