+++ cmd/cosign/cli/options/attestblob.go
@@ -27,9 +27,7 @@ type AttestBlobOptions struct {
 	Cert      string
 	CertChain string
 	NoUpload  bool
-	Force     bool
 	Recursive bool
-	Replace   bool
 	Timeout   time.Duration
 	Hash      string
 
@@ -64,12 +62,6 @@ func (o *AttestBlobOptions) AddFlags(cmd *cobra.Command) {
 
 	cmd.Flags().BoolVar(&o.NoUpload, "no-upload", false,
 		"do not upload the generated attestation")
-
-	cmd.Flags().BoolVarP(&o.Force, "force", "f", false,
-		"skip warnings and confirmations")
-
-	cmd.Flags().BoolVarP(&o.Replace, "replace", "", false,
-		"")
 
 	cmd.Flags().DurationVar(&o.Timeout, "timeout", time.Second*30,
 		"HTTP Timeout defaults to 30 seconds")
